Fixed graph download to suggest file name
When clicking 'Save Graph' a series of fallbacks is
now tried.  The first and most reliable method is to send
the graph to a server which serves it back with a suggested
file name.  We fall back to a datauri if the server appears
to be malfunctioning (e.g., we are running locally off the filesystem)